# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/workstation_losses:
    get:
      tags:
        - WorkstationLosses
      summary: List WorkstationLosses
      description: Get a paginated list of WorkstationLosses with optional filtering and sorting
      operationId: listWorkstationLosses
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkstationLossList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - WorkstationLosses
      summary: Create WorkstationLoss
      description: Create a new WorkstationLoss
      operationId: createWorkstationLoss
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkstationLossRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkstationLoss'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/workstation_productivities:
    get:
      tags:
        - WorkstationProductivities
      summary: List WorkstationProductivities
      description: Get a paginated list of WorkstationProductivities with optional filtering and sorting
      operationId: listWorkstationProductivities
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkstationProductivityList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - WorkstationProductivities
      summary: Create WorkstationProductivity
      description: Create a new WorkstationProductivity
      operationId: createWorkstationProductivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkstationProductivityRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkstationProductivity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    WorkstationLoss:
      type: object
      required:
        - id
        - name
        - loss_type
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        loss_type:
          type: string
        metadata:
          type: object
...
    CreateWorkstationLossRequest:
      type: object
      required:
        - name
        - loss_type
        - metadata
      properties:
        company_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
          maxLength: 140
        loss_type:
          type: string
        metadata:
          type: object
...
    WorkstationLossList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkstationLoss'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    WorkstationProductivity:
      type: object
      required:
        - id
        - company_id
        - workstation_id
        - loss_id
        - date_start
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        workstation_id:
          type: string
          format: uuid
        job_card_id:
          nullable: true
          type: string
          format: uuid
        loss_id:
          type: string
          format: uuid
        date_start:
          type: string
          format: date-time
        date_end:
          nullable: true
          type: string
          format: date-time
        description:
          nullable: true
          type: string
        metadata:
          type: object
...
    CreateWorkstationProductivityRequest:
      type: object
      required:
        - company_id
        - workstation_id
        - loss_id
        - date_start
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        workstation_id:
          type: string
          format: uuid
        job_card_id:
          nullable: true
          type: string
          format: uuid
        loss_id:
          type: string
          format: uuid
        date_start:
          type: string
          format: date-time
        date_end:
          nullable: true
          type: string
          format: date-time
        description:
          nullable: true
          type: string
        metadata:
          type: object
...
    WorkstationProductivityList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkstationProductivity'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []